Fixed issue with sieveslit fit in which the inclusion of non-dispersive angle made the fit worse. The issue was with errors that existed in the independent variables not being accounted for. This is a difficult statistical issue known as Errors-in-variables:
In statistics, errors-in-variables models or measurement error models
are regression models that account for measurement errors in the
independent variables. In contrast, standard regression models
assume that those regressors have been measured exactly,
or observed without error; as such, those models account
only for errors in the dependent variables, or responses.
-Wikipedia https://en.wikipedia.org/wiki/Errors-in-variables_models

The solution to the issue was to artificially add extra data points in the sieveslit (x,a,y,b) data, which was representative of
data 1sigma out in y and 1sigma out in b. Then the fit tries to bring all points to the center of a given hole. It obviously cannot
do so, but by doing this, information about error in the centroids is incorporated in an intermediate way.